+++ bugsnag-android-core/src/main/java/com/bugsnag/android/NativeInterface.java
@@ -90,6 +90,7 @@ public class NativeInterface {
         data.put("durationInForeground", app.getDurationInForeground());
         data.put("versionCode", app.getVersionCode());
         data.put("inForeground", app.getInForeground());
+        data.put("isLaunching", app.isLaunching());
         data.put("binaryArch", app.getBinaryArch());
         data.putAll(source.getAppDataMetadata());
         return data;
@@ -306,11 +307,13 @@ public class NativeInterface {
      *                          stages
      * @param payloadBytes The raw JSON payload of the event
      * @param apiKey The apiKey for the event
+     * @param isLaunching whether the crash occurred when the app was launching
      */
     @SuppressWarnings("unused")
     public static void deliverReport(@Nullable byte[] releaseStageBytes,
                                      @NonNull byte[] payloadBytes,
-                                     @NonNull String apiKey) {
+                                     @NonNull String apiKey,
+                                     boolean isLaunching) {
         if (payloadBytes == null) {
             return;
         }
@@ -326,8 +329,10 @@ public class NativeInterface {
             EventStore eventStore = client.getEventStore();
 
             String filename = eventStore.getNdkFilename(payload, apiKey);
+            if (isLaunching) {
+                filename = filename.replace(".json", "startupcrash.json");
+            }
             eventStore.enqueueContentForDelivery(payload, filename);
-            eventStore.flushAsync();
         }
     }