+++ bugsnag-android-core/src/test/java/com/bugsnag/android/ConfigurationFacadeTest.java
@@ -91,16 +91,29 @@ public class ConfigurationFacadeTest {
         assertTrue(config.impl.getPersistUser());
     }
 
+    @SuppressWarnings("deprecation")
     @Test
     public void launchCrashThresholdMsValid() {
         config.setLaunchCrashThresholdMs(123456);
-        assertEquals(123456, config.impl.getLaunchCrashThresholdMs());
+        assertEquals(123456, config.impl.getLaunchDurationMillis());
+    }
+
+    @Test
+    public void launchDurationMillisValid() {
+        config.setLaunchDurationMillis(123456);
+        assertEquals(123456, config.impl.getLaunchDurationMillis());
     }
 
     @Test
     public void autoTrackSessionsValid() {
         config.setAutoTrackSessions(true);
         assertTrue(config.impl.getAutoTrackSessions());
+    }
+
+    @Test
+    public void sendLaunchCrashesSynchronouslyValid() {
+        config.setSendLaunchCrashesSynchronously(true);
+        assertTrue(config.impl.getSendLaunchCrashesSynchronously());
     }
 
     @Test