+++ bugsnag-android-core/src/test/java/com/bugsnag/android/ConfigurationTest.java
@@ -94,8 +94,9 @@ public class ConfigurationTest {
         return BugsnagTestUtils.convert(config);
     }
 
+    @SuppressWarnings("deprecation")
     @Test
-    public void testLaunchThreshold() {
+    public void testLaunchThresholdDeprecated() {
         assertEquals(5000L, config.getLaunchCrashThresholdMs());
 
         config.setLaunchCrashThresholdMs(-5);
@@ -104,6 +105,18 @@ public class ConfigurationTest {
         int expected = 1500;
         config.setLaunchCrashThresholdMs(expected);
         assertEquals(expected, config.getLaunchCrashThresholdMs());
+    }
+
+    @Test
+    public void testLaunchThreshold() {
+        assertEquals(5000L, config.getLaunchDurationMillis());
+
+        config.setLaunchDurationMillis(-5);
+        assertEquals(5000, config.getLaunchDurationMillis());
+
+        int expected = 1500;
+        config.setLaunchDurationMillis(expected);
+        assertEquals(expected, config.getLaunchDurationMillis());
     }
 
     @Test