+++ CHANGELOG.md
@@ -2,11 +2,28 @@
 
 ## TBD
 
+### Deprecations
+
+* `Configuration#launchCrashThresholdMs` is deprecated in favour of `Configuration#launchDurationMillis`
+
 ### Enhancements
 
 * Improve detection of rooted devices
   [#1194](https://github.com/bugsnag/bugsnag-android/pull/1194)
   [#1195](https://github.com/bugsnag/bugsnag-android/pull/1195)
+
+* Add public API for crash-on-launch detection
+  [#1157](https://github.com/bugsnag/bugsnag-android/pull/1157)
+  [#1159](https://github.com/bugsnag/bugsnag-android/pull/1159)
+  [#1165](https://github.com/bugsnag/bugsnag-android/pull/1165)
+  [#1164](https://github.com/bugsnag/bugsnag-android/pull/1164)
+  [#1182](https://github.com/bugsnag/bugsnag-android/pull/1182)
+  [#1184](https://github.com/bugsnag/bugsnag-android/pull/1184)
+  [#1185](https://github.com/bugsnag/bugsnag-android/pull/1185)
+  [#1186](https://github.com/bugsnag/bugsnag-android/pull/1186)
+  [#1180](https://github.com/bugsnag/bugsnag-android/pull/1180)
+  [#1188](https://github.com/bugsnag/bugsnag-android/pull/1188)
+  [#1191](https://github.com/bugsnag/bugsnag-android/pull/1191)
 
 ## 5.7.1 (2021-03-03)